requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
passlib
psycopg2-binary
